Update in the DNA table
Update in the DNA table of the samples located in /Namibia_project/Data/Nanodrop_measurements: I performed additional DNA cleaning because in our meeting we wanted to see if there are still contaminants in the samples that caused a division in the samples after the 16s gene sequencing (some samples showed a rapid plateau at 20 species, while others did not). Here in the table, I included Nanodrop and Quibit values before and after the extra cleaning. The cleaning improved the 260/230 ratio in several samples, but not the 260/280.